main/requirements.txt
Flask==0.10.1
Flask-Admin==1.4.0
Flask-SQLAlchemy==2.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.11
WTForms==2.1
Werkzeug==0.11.4
argparse==1.2.1
itsdangerous==0.24
python-dateutil==2.4.2
requests==2.9.1
six==1.10.0
wsgiref==0.1.2
psycopg2==2.7.1
schedule==0.3.2
gunicorn==19.4.5
html5lib==1.0b8
xlrd==1.0.0
